pyproject.toml
[build-system] # https://python-poetry.org/docs/pyproject/#poetry-and-pep-517
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry] # https://python-poetry.org/docs/pyproject/
name = "dslmodel"
version = "2024.11.19"
description = "Pydantic + DSPy instances from prompts and Jinja."
authors = ["Sean Chatman <info@chatmangpt.com>"]
readme = "README.md"
repository = "https://github.com/seanchatmangpt/dslmodel"

[tool.poetry.scripts] # https://python-poetry.org/docs/pyproject/#scripts
dsl = "dslmodel.cli:app"
jygen = "jygen.cli:app"

[tool.poetry.dependencies] # https://python-poetry.org/docs/dependency-specification/
coloredlogs = ">=15.0.1"
poethepoet = ">=0.25.0"
python = ">=3.12,<4.0"
dspy-ai = "^2.5.6"
pyperclip = "^1.9.0"
inflection = "^0.5.1"
jinja2-time = "^0.2.0"
jinja2-ext = "^0.1"
loguru = "^0.7.2"
munch = "^4.0.0"
ijson = "^3.3.0"
factory-boy = "^3.3.1"
inject = "^5.2.1"
pytest-xdist = "^3.6.1"
jupyter = "^1.1.1"
wcmatch = "^10.0"
pandasql = "^0.7.3"
pytest = "^8.3.3"
transitions = "^0.9.2"
apscheduler = "^3.10.4"
pygithub = "^2.4.0"
hypothesis = "^6.114.1"
reactivex = "^4.0.4"
pyd4all = "^2024.10.29"
toml = "^0.10.2"
python-socketio = { extras = ["asyncio-client"], version = "^5.11.4" }
injector = "^0.22.0"
mcp = "^1.0.0"
pydantic-ai = {extras = ["logfire"], version = "^0.0.8"}
nest-asyncio = "^1.6.0"

[tool.poetry.group.test.dependencies] # https://python-poetry.org/docs/master/managing-dependencies/
coverage = { extras = ["toml"], version = ">=7.4.4" }
pytest = ">=8.1.1"
pytest-mock = ">=3.14.0"
pytest-xdist = ">=3.5.0"
pytest-watch = ">=4.2.0"


[tool.poetry.group.dev.dependencies] # https://python-poetry.org/docs/master/managing-dependencies/
cruft = ">=2.15.0"
ipykernel = ">=6.29.4"
ipython = ">=8.23.0"
ipywidgets = ">=8.1.2"
pdoc = ">=14.4.0"
pytest-asyncio = "^0.24.0"

[tool.coverage.report] # https://coverage.readthedocs.io/en/latest/config.html#report
precision = 1
show_missing = true
skip_covered = true

[tool.coverage.run] # https://coverage.readthedocs.io/en/latest/config.html#run
branch = true
command_line = "--module pytest"
data_file = "reports/.coverage"
source = ["src"]

[tool.coverage.xml] # https://coverage.readthedocs.io/en/latest/config.html#xml
output = "reports/coverage.xml"

#[tool.mypy]  # https://mypy.readthedocs.io/en/latest/config_file.html
#junit_xml = "reports/mypy.xml"
#plugins = "pydantic.mypy"
#ignore_missing_imports = true
#pretty = true
#show_column_numbers = true
#show_error_codes = true
#show_error_context = true
#warn_unreachable = true

[tool.pytest.ini_options] # https://docs.pytest.org/en/latest/reference/reference.html#ini-options-ref
addopts = "--color=yes --doctest-modules --exitfirst --failed-first --verbosity=2 --junitxml=reports/pytest.xml --asyncio-mode=auto"
testpaths = ["src", "tests"]
xfail_strict = true

[tool.poe.tasks] # https://github.com/nat-n/poethepoet
push = "act push --container-architecture linux/amd64 --secret-file .secrets"
pub = "poetry publish --build"
pipe = "pip install -e ."


[tool.poe.tasks.api]
help = "Serve the REST API"
shell = """
    if [ $dev ]
    then {
      uvicorn \
        --host $host \
        --port $port \
        --reload \
        dslmodel.api:app
    } else {
      gunicorn \
        --access-logfile - \
        --bind $host:$port \
        --graceful-timeout 10 \
        --keep-alive 10 \
        --log-file - \
        --timeout 30 \
        --worker-class uvicorn.workers.UvicornWorker \
        --worker-tmp-dir /dev/shm \
        --workers 2 \
        dslmodel.api:app
    } fi
    """

[[tool.poe.tasks.api.args]]
help = "Bind socket to this host (default: 0.0.0.0)"
name = "host"
options = ["--host"]
default = "0.0.0.0"

[[tool.poe.tasks.api.args]]
help = "Bind socket to this port (default: 8000)"
name = "port"
options = ["--port"]
default = "8000"

[[tool.poe.tasks.api.args]]
help = "Enable development mode"
type = "boolean"
name = "dev"
options = ["--dev"]

[tool.poe.tasks.docs]
help = "Generate this app's docs"
cmd = """
    pdoc
      --docformat $docformat
      --output-directory $outputdirectory
      dslmodel
    """

[[tool.poe.tasks.docs.args]]
help = "The docstring style (default: numpy)"
name = "docformat"
options = ["--docformat"]
default = "numpy"

[[tool.poe.tasks.docs.args]]
help = "The output directory (default: docs)"
name = "outputdirectory"
options = ["--output-directory"]
default = "docs"

[tool.poe.tasks.test]
help = "Test this app"

[[tool.poe.tasks.test.sequence]]
cmd = "coverage run"

[[tool.poe.tasks.test.sequence]]
cmd = "coverage report"

[[tool.poe.tasks.test.sequence]]
cmd = "coverage xml"

[tool.poe.tasks.test-watch]
help = "Watch for changes and rerun tests"
cmd = """poetry run ptw --onfail="notify-send 'Tests failed!'" --clear --runner="pytest --tb=short --maxfail=1 --color=yes --doctest-modules --junitxml=reports/pytest.xml --failed-first --exitfirst --verbosity=2" """
